Split up chart components
